# backend/docs/optimization/api/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Prompt Card System Optimization API
  version: 1.0.0
  description: |
    Comprehensive API for the Prompt Card System optimization strategies including:
    - AdvancedKVCache with MorphKV adaptive optimization
    - RealTimeOptimizer with ML-driven feedback loops  
    - EdgeOptimizer with distributed computing architecture
    
    This API provides endpoints for optimization management, performance monitoring,
    and real-time analytics for production deployment.
    
  contact:
    name: Prompt Card System Team
    email: support@promptcard.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.promptcard.dev/v1
    description: Production server
  - url: https://staging-api.promptcard.dev/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /optimization/edge/nodes:
    get:
      summary: List edge nodes
      description: Get list of all registered edge nodes
      tags:
        - Edge Computing
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, online, offline]
          description: Filter nodes by status
          example: "online"
        - name: region
          in: query
          required: false
          schema:
            type: string
          description: Filter nodes by region
          example: "us-west"
      responses:
        '200':
          description: Edge nodes list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EdgeNode'

    post:
      summary: Register edge node
      description: Register a new edge node in the distributed system
      tags:
        - Edge Computing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeNodeRegistration'
      responses:
        '201':
          description: Edge node registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EdgeNodeRegistrationResult'
  /optimization/edge/nodes/{nodeId}:
    get:
      summary: Get edge node details
      description: Get detailed information about a specific edge node
      tags:
        - Edge Computing
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
          description: Edge node identifier
      responses:
        '200':
          description: Edge node details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EdgeNode'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Remove edge node
      description: Remove an edge node from the distributed system
      tags:
        - Edge Computing
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
          description: Edge node identifier
      responses:
        '200':
          description: Edge node removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EdgeNode:
      type: object
      properties:
        id:
          type: string
          example: "edge-us-west-1"
        location:
          type: object
          properties:
            region:
              type: string
              example: "us-west"
            city:
              type: string
              example: "San Francisco"
            country:
              type: string
              example: "US"
            latitude:
              type: number
              example: 37.7749
            longitude:
              type: number
              example: -122.4194
            timezone:
              type: string
              example: "America/Los_Angeles"
        capabilities:
          type: object
          properties:
            prompt_optimization:
              type: boolean
              example: true
            semantic_analysis:
              type: boolean
              example: true
            model_inference:
              type: boolean
              example: true
            vector_search:
              type: boolean
              example: true
            caching:
              type: boolean
              example: true
            compression:
              type: boolean
              example: true
            load_balancing:
              type: boolean
              example: true
        resources:
          type: object
          properties:
            cpu_cores:
              type: integer
              example: 8
            memory_gb:
              type: integer
              example: 32
            storage_gb:
              type: integer
              example: 500
            network_mbps:
              type: integer
              example: 1000
            gpu_count:
              type: integer
              example: 2
            gpu_memory_gb:
              type: integer
              example: 16
        status:
          type: object
          properties:
            online:
              type: boolean
              example: true
            last_heartbeat:
              type: string
              format: date-time
            current_load:
              type: number
              minimum: 0
              maximum: 1
              example: 0.35
            health_score:
              type: number
              minimum: 0
              maximum: 100
              example: 92
    EdgeNodeRegistration:
      type: object
      required:
        - id
        - location
        - capabilities
        - resources
      properties:
        id:
          type: string
          example: "edge-us-west-1"
        location:
          type: object
          required:
            - region
            - city
            - country
            - latitude
            - longitude
            - timezone
          properties:
            region:
              type: string
              example: "us-west"
            city:
              type: string
              example: "San Francisco"
            country:
              type: string
              example: "US"
            latitude:
              type: number
              example: 37.7749
            longitude:
              type: number
              example: -122.4194
            timezone:
              type: string
              example: "America/Los_Angeles"
        capabilities:
          type: object
          properties:
            prompt_optimization:
              type: boolean
              example: true
            semantic_analysis:
              type: boolean
              example: true
            model_inference:
              type: boolean
              example: false
            vector_search:
              type: boolean
              example: true
            caching:
              type: boolean
              example: true
            compression:
              type: boolean
              example: true
            load_balancing:
              type: boolean
              example: true
        resources:
          type: object
          required:
            - cpu_cores
            - memory_gb
            - storage_gb
            - network_mbps
          properties:
            cpu_cores:
              type: integer
              minimum: 2
              example: 8
            memory_gb:
              type: integer
              minimum: 4
              example: 32
            storage_gb:
              type: integer
              minimum: 50
              example: 500
            network_mbps:
              type: integer
              minimum: 100
              example: 1000
            gpu_count:
              type: integer
              minimum: 0
              example: 2
            gpu_memory_gb:
              type: integer
              minimum: 0
              example: 16
    EdgeNodeRegistrationResult:
      type: object
      properties:
        success:
          type: boolean
          example: true
        node_id:
          type: string
          example: "edge-us-west-1"
        deployment_time_ms:
          type: number
          example: 342.5
        initial_health_score:
          type: number
          example: 95
        assigned_workloads:
          type: array
          items:
            type: string
          example: []
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        details:
          type: string
          example: "Missing required field: promptId"
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Resource not found"
            details: "The requested resource could not be found"
            timestamp: "2024-01-01T12:00:00Z"
...